pyproject.toml
[project]
name = "recommendation-system"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Leonardo Lazzaro", email = "llazzaro@dc.uba.ar" }
]
dependencies = [
    "django>=5.0.2",
    "djangorestframework>=3.14.0",
    "factory-boy>=3.3.0",
    "django-extensions>=3.2.3",
]
readme = "README.md"
requires-python = ">= 3.12"

[project.scripts]
hello = "recommendation_system:hello"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.isort]
profile = "black"

[tool.rye]
managed = true
dev-dependencies = [
    "black>=24.2.0",
    "mypy>=1.8.0",
    "pytest>=8.0.2",
    "pytest-django>=4.8.0",
    "factory-boy>=3.3.0",
    "isort>=5.13.2",
    "ruff>=0.3.4",
    "django-stub>=0.1",
    "django-stubs>=4.2.7",
    "ipdb>=0.13.13",
    "requests>=2.31.0",
    "tqdm>=4.66.2",
    "types-requests>=2.31.0.20240311",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/recommendation_system"]
